Add a debug bytecode disassembler template class for compiler and VM
Make the debug disassemblers in the VM and compiler use a common
implementation.
