+++ StdAfx.h
@@ -2,30 +2,6 @@
 //      or project specific include files that are used frequently,
 //      but are changed infrequently
 
-#if !defined(AFX_STDAFX_H__AB93FFD3_2625_11D3_8362_FC91FD000000__INCLUDED_)
-#define AFX_STDAFX_H__AB93FFD3_2625_11D3_8362_FC91FD000000__INCLUDED_
-
-#if _MSC_VER > 1000
 #pragma once
-#endif // _MSC_VER > 1000
 
-#include "segdefs.h"
-
-#pragma code_seg(ATL_SEG)
-
-#ifndef STRICT
-	#define STRICT
-#endif
-
-#define _ATL_APARTMENT_THREADED
-
-#define _ATL_ALL_WARNINGS
-#include <atlbase.h>
-#include <atlcom.h>
-
-#pragma code_seg()
-
-//{{AFX_INSERT_LOCATION}}
-// Microsoft Visual C++ will insert additional declarations immediately before the previous line.
-
-#endif // !defined(AFX_STDAFX_H__AB93FFD3_2625_11D3_8362_FC91FD000000__INCLUDED)
+#include "ist.h"
